move create_app() out of __init__.py
